backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-multipart==0.0.9
pandas==2.2.0
openpyxl==3.1.2
pyyaml==6.0.1
anthropic==0.40.0
google-genai>=1.0.0
httpx>=0.27.0
firebase-admin>=7.0.0
lightgbm>=4.0.0
scikit-learn>=1.3.0
